+++ src/Components/BarAndLineChartsWrapper/renderLineInBarChart/renderDataPoints.tsx
@@ -3,8 +3,9 @@ import {styles} from '../../../BarChart/styles';
 import {View} from 'react-native';
 import {getXForLineInBar, getYForLineInBar} from 'gifted-charts-core';
 import {Rect, Text as CanvasText, Circle} from 'react-native-svg';
+import {DataPointProps} from 'gifted-charts-core';
 
-export const renderDataPoints = (props: any) => {
+export const renderDataPoints = (props: DataPointProps) => {
   const {
     data,
     lineConfig,
@@ -14,13 +15,29 @@ export const renderDataPoints = (props: any) => {
     firstBarWidth,
     yAxisLabelWidth,
     spacing,
+    selectedIndex,
   } = props;
   return data.map((item: any, index: number) => {
-    if (index < lineConfig.startIndex || index > lineConfig.endIndex || item.hideDataPoint) {
+    if (
+      index < lineConfig.startIndex ||
+      index > lineConfig.endIndex ||
+      item.hideDataPoint
+    ) {
       return null;
     }
     const currentBarWidth = item.barWidth || barWidth || 30;
     const customDataPoint = item.customDataPoint || lineConfig.customDataPoint;
+    const dataPointColor =
+      lineConfig.focusEnabled &&
+      index === (lineConfig.focusedDataPointIndex ?? selectedIndex)
+        ? lineConfig.focusedDataPointColor
+        : lineConfig.dataPointsColor;
+
+    const dataPointRadius =
+      lineConfig.focusEnabled &&
+      index === (lineConfig.focusedDataPointIndex ?? selectedIndex)
+        ? lineConfig.focusedDataPointRadius
+        : lineConfig.dataPointsRadius;
     const value =
       item.value ??
       item.stacks.reduce((total: number, item: any) => total + item.value, 0);
@@ -73,7 +90,7 @@ export const renderDataPoints = (props: any) => {
             }
             width={lineConfig.dataPointsWidth}
             height={lineConfig.dataPointsHeight}
-            fill={lineConfig.dataPointsColor}
+            fill={dataPointColor}
           />
           {item.dataPointText && (
             <CanvasText
@@ -122,8 +139,8 @@ export const renderDataPoints = (props: any) => {
             containerHeight,
             maxValue,
           )}
-          r={lineConfig.dataPointsRadius}
-          fill={lineConfig.dataPointsColor}
+          r={dataPointRadius}
+          fill={dataPointColor}
         />
         {item.dataPointText && (
           <CanvasText